Move new build-coordinator to remote-build-coordinator module.

+++ facade/src/main/java/org/jboss/pnc/facade/providers/api/SCMRepositoryProvider.java
@@ -90,7 +90,7 @@ public interface SCMRepositoryProvider
 
     @Data
     public static class RepositoryCreated {
-        private final Integer taskId;
+        private final Long taskId;
         private final int repositoryId;
     }
 }
